+++ mobile/lib/services/game_area_service.dart
@@ -25,7 +25,7 @@ class GameAreaService extends ChangeNotifier {
   // Par défaut la vitesse c'est SPEED_X1 si tu call showDifferenceFound
   // avec seulement des coordonnées, même logique pour toggleCheatMode et showDifferenceNotFound
   void showDifferenceFound(List<Coordinate> newCoordinates,
-      [int flashingSpeed = SPEED_X1]) {
+      [double flashingSpeed = SPEED_X1]) {
     if (newCoordinates.isNotEmpty) {
       soundService.playCorrectSound();
       coordinates.addAll(newCoordinates);
@@ -40,7 +40,7 @@ class GameAreaService extends ChangeNotifier {
   }
 
   void showDifferenceNotFound(Coordinate currentCoord, bool isLeft,
-      [int flashingSpeed = SPEED_X1]) {
+      [double flashingSpeed = SPEED_X1]) {
     if (isLeft) {
       isClickDisabled = true;
       soundService.playErrorSound();
@@ -78,7 +78,7 @@ class GameAreaService extends ChangeNotifier {
   }
 
   Future<void> startBlinking(List<Coordinate> coords,
-      [int flashingSpeed = SPEED_X1]) async {
+      [double flashingSpeed = SPEED_X1]) async {
     initPath(coords);
     if (blinkingDifference == null) return;
 
@@ -128,7 +128,7 @@ class GameAreaService extends ChangeNotifier {
   }
 
   Future<void> toggleCheatMode(List<Coordinate> coords,
-      [int flashingSpeed = SPEED_X1]) async {
+      [double flashingSpeed = SPEED_X1]) async {
     isCheatMode = !isCheatMode;
     if (isCheatMode) {
       initCheatPath(coords);
